test/fix: e2e swap quotes test in one for zero direction (right); some bug fixes (#1773)
* WIP: e2e swap quotes test in one for zero direction (right)

* WIP: e2e swap quotes test in one for zero direction (right)

* continue fixing edge case around low liquidity error

* test/fix: more descriptive error message when failing quotes due to lack of precision

* updates

* update tests

* update tests

* update tests

* remove prints

* reset change

* reset more

* updates

* updates

* updates

* update

* fix add to position slippage

* build

---------

+++ packages/stores/src/account/osmosis/index.ts
@@ -631,24 +631,63 @@ export class OsmosisAccountImpl {
           .filter((coin): coin is Coin => coin !== undefined)
           .sort((a, b) => a?.denom.localeCompare(b?.denom))
           .map(({ denom, amount }) => ({ denom, amount: amount.toString() }));
-        const token_min_amount0 =
-          baseCoin &&
-          // full tolerance if 0 sqrt price so no positions
+
+        // full tolerance if 0 sqrt price so no positions
+        let token_min_amount0 = "0";
+        let token_min_amount1 = "0";
+
+        // 3 cases:
+        // - If position is active, consists of both tokens
+        // - If position is under current tick, consists only of token 1.
+        // - If position is above current tick, consists only of token 0.
+        if (
           !queryPool.concentratedLiquidityPoolInfo?.currentSqrtPrice.isZero()
-            ? new Dec(baseCoin.amount)
-                .mul(new Dec(1).sub(new Dec(maxSlippage).quo(new Dec(100))))
-                .truncate()
-                .toString()
-            : "0";
-        const token_min_amount1 =
-          quoteCoin &&
-          // full tolerance if 0 sqrt price so no positions
-          !queryPool.concentratedLiquidityPoolInfo?.currentSqrtPrice.isZero()
-            ? new Dec(quoteCoin.amount)
-                .mul(new Dec(1).sub(new Dec(maxSlippage).quo(new Dec(100))))
-                .truncate()
-                .toString()
-            : "0";
+        ) {
+          const currentSqrtPrice =
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion, @typescript-eslint/no-non-null-asserted-optional-chain
+            queryPool.concentratedLiquidityPoolInfo?.currentSqrtPrice!;
+
+          const currentTick = OsmosisMath.priceToTick(
+            currentSqrtPrice.mul(currentSqrtPrice)
+          );
+
+          const slippageMultiplier = new Dec(1).sub(
+            new Dec(maxSlippage).quo(new Dec(100))
+          );
+
+          if (currentTick >= lowerTick && currentTick < upperTick) {
+            // Position consists of both tokens
+            token_min_amount0 = baseCoin
+              ? new Dec(baseCoin.amount)
+                  .mul(slippageMultiplier)
+                  .truncate()
+                  .toString()
+              : token_min_amount0;
+
+            token_min_amount1 = quoteCoin
+              ? new Dec(quoteCoin.amount)
+                  .mul(slippageMultiplier)
+                  .truncate()
+                  .toString()
+              : token_min_amount1;
+          } else if (currentTick < lowerTick) {
+            // Position consists of 1 token only.
+            token_min_amount0 = baseCoin
+              ? new Dec(baseCoin.amount)
+                  .mul(slippageMultiplier)
+                  .truncate()
+                  .toString()
+              : token_min_amount0;
+          } else if (currentTick >= upperTick) {
+            // Position consists of 1 token only.
+            token_min_amount1 = quoteCoin
+              ? new Dec(quoteCoin.amount)
+                  .mul(slippageMultiplier)
+                  .truncate()
+                  .toString()
+              : token_min_amount1;
+          }
+        }
 
         const msg = this.msgOpts.clCreatePosition.messageComposer({
           poolId: BigInt(poolId),
